moving models into separate pages.py and django.py files
