+++ teachers_digital_platform/models/__init__.py
@@ -1,0 +1,8 @@
+from teachers_digital_platform.models.django import (
+    BaseActivityTaxonomy, ActivityBuildingBlock, ActivitySchoolSubject, ActivityTopic, ActivityGradeLevel, ActivityAgeRange,
+    ActivitySpecialPopulation, ActivityType, ActivityTeachingStrategy, ActivityBloomsTaxonomyLevel,
+    ActivityDuration, ActivityJumpStartCoalition, ActivityCouncilForEconEd
+)
+from teachers_digital_platform.models.pages import (
+    ActivityIndexPage, ActivityPage
+)
